ProtectedBranchPolicy used from Controller for destroy/update

+++ app/services/protected_branches/destroy_service.rb
@@ -1,6 +1,8 @@
 module ProtectedBranches
   class DestroyService < BaseService
     def execute(protected_branch)
+      raise Gitlab::Access::AccessDeniedError unless can?(current_user, :destroy_protected_branch, protected_branch)
+
       protected_branch.destroy
     end
   end
